+++ nile-stand/include/timing.h
@@ -32,7 +32,7 @@ uint64_t timing_time_since_us(timing_marker_t marker);
 /**
  * Time since the given marking in seconds.
  */
-double timing_time_since_s(timing_marker_t marker);
+float timing_time_since_s(timing_marker_t marker);
 
 /**
  * Gets the main loop's delta time in microseconds.
@@ -42,12 +42,12 @@ uint64_t timing_delta_time_us(void);
 /**
  * Gets the main loop's delta time in seconds.
  */
-double timing_delta_time_s(void);
+float timing_delta_time_s(void);
 
 /**
  * Returns the rate of change over time (in seconds) of the given initial and
  * final values.
  */
-double timing_d_dt(double x0, double x1);
+float timing_d_dt(float x0, float x1);
 
 #endif
